Add elimination tree computation.
