+++ src/sparse/mod.rs
@@ -1,3 +1,8 @@
-pub use self::cs_matrix::{CsMatrix, CsVector};
+pub use self::cs_matrix::{CsMatrix, CsStorage, CsStorageMut, CsVector};
 
 mod cs_matrix;
+mod cs_matrix_analysis;
+mod cs_matrix_cholesky;
+mod cs_matrix_conversion;
+mod cs_matrix_ops;
+mod cs_matrix_solve;
